requirements.txt
Flask==3.1.0
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
pyzbar==0.1.9
qrcode==8.0
Werkzeug==3.1.3
Gunicorn
openpyxl
pandas
